requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
azure-identity
azure-ai-inference
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
tiktoken
